+++ backend/app/controllers/spree/admin/products_controller.rb
@@ -1,8 +1,6 @@
 module Spree
   module Admin
     class ProductsController < ResourceController
-      helper 'spree/products'
-
       before_filter :load_data, :except => [:index]
       create.before :create_before
       update.before :update_before
