+++ core/app/helpers/spree/base_helper.rb
@@ -1,120 +1,8 @@
 module Spree
   module BaseHelper
-
-    # Defined because Rails' current_page? helper is not working when Spree is mounted at root.
-    def current_spree_page?(url)
-      path = request.fullpath.gsub(/^\/\//, '/')
-      if url.is_a?(String)
-        return path == url
-      elsif url.is_a?(Hash)
-        return path == spree.url_for(url)
-      end
-      return false
-    end
-
-    def link_to_cart(text = nil)
-      text = text ? h(text) : Spree.t(:cart)
-      css_class = nil
-
-      if simple_current_order.nil? or simple_current_order.item_count.zero?
-        text = "#{text}: (#{Spree.t(:empty)})"
-        css_class = 'empty'
-      else
-        text = "#{text}: (#{simple_current_order.item_count})  <span class='amount'>#{simple_current_order.display_total.to_html}</span>"
-        css_class = 'full'
-      end
-
-      link_to text.html_safe, spree.cart_path, :class => "cart-info #{css_class}"
-    end
-
     # human readable list of variant options
     def variant_options(v, options={})
       v.options_text
-    end
-
-    def meta_data
-      object = instance_variable_get('@'+controller_name.singularize)
-      meta = {}
-
-      if object.kind_of? ActiveRecord::Base
-        meta[:keywords] = object.meta_keywords if object[:meta_keywords].present?
-        meta[:description] = object.meta_description if object[:meta_description].present?
-      end
-
-      if meta[:description].blank? && object.kind_of?(Spree::Product)
-        meta[:description] = strip_tags(truncate(object.description, length: 160, separator: ' '))
-      end
-
-      meta.reverse_merge!({
-        keywords: current_store.meta_keywords,
-        description: current_store.meta_description,
-      }) if meta[:keywords].blank? or meta[:description].blank?
-      meta
-    end
-
-    def meta_data_tags
-      meta_data.map do |name, content|
-        tag('meta', name: name, content: content)
-      end.join("\n")
-    end
-
-    def body_class
-      @body_class ||= content_for?(:sidebar) ? 'two-col' : 'one-col'
-      @body_class
-    end
-
-    def logo(image_path=Spree::Config[:logo])
-      link_to image_tag(image_path), spree.root_path
-    end
-
-    def flash_messages(opts = {})
-      ignore_types = ["order_completed"].concat(Array(opts[:ignore_types]).map(&:to_s) || [])
-
-      flash.each do |msg_type, text|
-        unless ignore_types.include?(msg_type)
-          concat(content_tag :div, text, class: "flash #{msg_type}")
-        end
-      end
-      nil
-    end
-
-    def breadcrumbs(taxon, separator="&nbsp;&raquo;&nbsp;", breadcrumb_class="inline")
-      return "" if current_page?("/") || taxon.nil?
-
-      crumbs = [[Spree.t(:home), spree.root_path]]
-
-      if taxon
-        crumbs << [Spree.t(:products), products_path]
-        crumbs += taxon.ancestors.collect { |a| [a.name, spree.nested_taxons_path(a.permalink)] } unless taxon.ancestors.empty?
-        crumbs << [taxon.name, spree.nested_taxons_path(taxon.permalink)]
-      else
-        crumbs << [Spree.t(:products), products_path]
-      end
-
-      separator = raw(separator)
-
-      crumbs.map! do |crumb|
-        content_tag(:li, itemscope:"itemscope", itemtype:"http://data-vocabulary.org/Breadcrumb") do
-          link_to(crumb.last, itemprop: "url") do
-            content_tag(:span, crumb.first, itemprop: "title")
-          end + (crumb == crumbs.last ? '' : separator)
-        end
-      end
-
-      content_tag(:nav, content_tag(:ul, raw(crumbs.map(&:mb_chars).join), class: breadcrumb_class), id: 'breadcrumbs', class: 'sixteen columns')
-    end
-
-    def taxons_tree(root_taxon, current_taxon, max_level = 1)
-      return '' if max_level < 1 || root_taxon.children.empty?
-      content_tag :ul, class: 'taxons-list' do
-        root_taxon.children.map do |taxon|
-          css_class = (current_taxon && current_taxon.self_and_ancestors.include?(taxon)) ? 'current' : nil
-          content_tag :li, class: css_class do
-           link_to(taxon.name, seo_url(taxon)) +
-           taxons_tree(taxon, current_taxon, max_level - 1)
-          end
-        end.join("\n").html_safe
-      end
     end
 
     def available_countries
@@ -130,18 +18,6 @@ module Spree
         country.name = Spree.t(country.iso, scope: 'country_names', default: country.name)
         country
       end.sort_by { |c| c.name.parameterize }
-    end
-
-    def seo_url(taxon)
-      return spree.nested_taxons_path(taxon.permalink)
-    end
-
-    def gem_available?(name)
-       Gem::Specification.find_by_name(name)
-    rescue Gem::LoadError
-       false
-    rescue
-       Gem.available?(name)
     end
 
     def display_price(product_or_variant)
@@ -162,13 +38,43 @@ module Spree
       end
     end
 
-    def link_to_tracking(shipment, options = {})
-      return unless shipment.tracking && shipment.shipping_method
-
-      if shipment.tracking_url
-        link_to(shipment.tracking, shipment.tracking_url, options)
+    # Returns the formatted price for the specified variant as a full price or
+    # a difference depending on configuration
+    #
+    # @param variant [Spree::Variant] the variant
+    # @return [Spree::Money] the price or price diff
+    def variant_price(variant)
+      if Spree::Config[:show_variant_full_price]
+        variant_full_price(variant)
       else
-        content_tag(:span, shipment.tracking)
+        variant_price_diff(variant)
+      end
+    end
+
+    # Returns the formatted price for the specified variant as a difference
+    # from product price
+    #
+    # @param variant [Spree::Variant] the variant
+    # @return [String] formatted string with label and amount
+    def variant_price_diff(variant)
+      variant_amount = variant.amount_in(current_currency)
+      product_amount = variant.product.amount_in(current_currency)
+      return if variant_amount == product_amount || product_amount.nil?
+      diff   = variant.amount_in(current_currency) - product_amount
+      amount = Spree::Money.new(diff.abs, currency: current_currency).to_html
+      label  = diff > 0 ? :add : :subtract
+      "(#{Spree.t(label)}: #{amount})".html_safe
+    end
+
+    # Returns the formatted full price for the variant, if at least one variant
+    # price differs from product price.
+    #
+    # @param variant [Spree::Variant] the variant
+    # @return [Spree::Money] the full price
+    def variant_full_price(variant)
+      product = variant.product
+      unless product.variants.active(current_currency).all? { |v| v.price == product.price }
+        Spree::Money.new(variant.price, { currency: current_currency }).to_html
       end
     end
