+++ core/spec/helpers/base_helper_spec.rb
@@ -3,7 +3,12 @@ require 'spec_helper'
 describe Spree::BaseHelper, :type => :helper do
   include Spree::BaseHelper
 
-  let(:current_store){ create :store }
+  let(:product) { create(:product) }
+  let(:currency) { 'USD' }
+
+  before do
+    allow(helper).to receive(:current_currency) { currency }
+  end
 
   context "available_countries" do
     let(:country) { create(:country) }
@@ -68,80 +73,6 @@ describe Spree::BaseHelper, :type => :helper do
     it "should raise NoMethodError when style is not exists" do
       expect { another_strange_image(product) }.to raise_error(NoMethodError)
     end
-
-  end
-
-  # Regression test for https://github.com/spree/spree/issues/2034
-  context "flash_message" do
-    let(:flash) { {"notice" => "ok", "foo" => "foo", "bar" => "bar"} }
-
-    it "should output all flash content" do
-      flash_messages
-      html = Nokogiri::HTML(helper.output_buffer)
-      expect(html.css(".notice").text).to eq("ok")
-      expect(html.css(".foo").text).to eq("foo")
-      expect(html.css(".bar").text).to eq("bar")
-    end
-
-    it "should output flash content except one key" do
-      flash_messages(:ignore_types => :bar)
-      html = Nokogiri::HTML(helper.output_buffer)
-      expect(html.css(".notice").text).to eq("ok")
-      expect(html.css(".foo").text).to eq("foo")
-      expect(html.css(".bar").text).to be_empty
-    end
-
-    it "should output flash content except some keys" do
-      flash_messages(:ignore_types => [:foo, :bar])
-      html = Nokogiri::HTML(helper.output_buffer)
-      expect(html.css(".notice").text).to eq("ok")
-      expect(html.css(".foo").text).to be_empty
-      expect(html.css(".bar").text).to be_empty
-      expect(helper.output_buffer).to eq("<div class=\"flash notice\">ok</div>")
-    end
-  end
-
-  context "link_to_tracking" do
-    it "returns tracking link if available" do
-      a = link_to_tracking_html(shipping_method: true, tracking: '123', tracking_url: 'http://g.c/?t=123').css('a')
-
-      expect(a.text).to eq '123'
-      expect(a.attr('href').value).to eq 'http://g.c/?t=123'
-    end
-
-    it "returns tracking without link if link unavailable" do
-      html = link_to_tracking_html(shipping_method: true, tracking: '123', tracking_url: nil)
-      expect(html.css('span').text).to eq '123'
-    end
-
-    it "returns nothing when no shipping method" do
-      html = link_to_tracking_html(shipping_method: nil, tracking: '123')
-      expect(html.css('span').text).to eq ''
-    end
-
-    it "returns nothing when no tracking" do
-      html = link_to_tracking_html(tracking: nil)
-      expect(html.css('span').text).to eq ''
-    end
-
-    def link_to_tracking_html(options = {})
-      node = link_to_tracking(double(:shipment, options))
-      Nokogiri::HTML(node.to_s)
-    end
-  end
-
-  # Regression test for https://github.com/spree/spree/issues/2396
-  context "meta_data_tags" do
-    it "truncates a product description to 160 characters" do
-      # Because the controller_name method returns "test"
-      # controller_name is used by this method to infer what it is supposed
-      # to be generating meta_data_tags for
-      text = Faker::Lorem.paragraphs(2).join(" ")
-      @test = Spree::Product.new(:description => text)
-      tags = Nokogiri::HTML.parse(meta_data_tags)
-      content = tags.css("meta[name=description]").first["content"]
-      assert content.length <= 160, "content length is not truncated to 160 characters"
-    end
   end
 
   # Regression test for https://github.com/spree/spree/issues/5384
@@ -168,6 +99,110 @@ describe Spree::BaseHelper, :type => :helper do
   context "pretty_time" do
     it "prints in a format" do
       expect(pretty_time(DateTime.new(2012, 5, 6, 13, 33))).to eq "May 06, 2012  1:33 PM"
+    end
+  end
+
+  context "#variant_price_diff" do
+    let(:product_price) { 10 }
+    let(:variant_price) { 10 }
+
+    before do
+      @variant = create(:variant, :product => product)
+      product.price = 15
+      @variant.price = 10
+      allow(product).to receive(:amount_in) { product_price }
+      allow(@variant).to receive(:amount_in) { variant_price }
+    end
+
+    subject { helper.variant_price(@variant) }
+
+    context "when variant is same as master" do
+      it { is_expected.to be_nil }
+    end
+
+    context "when the master has no price" do
+      let(:product_price) { nil }
+
+      it { is_expected.to be_nil }
+    end
+
+    context "when currency is default" do
+      context "when variant is more than master" do
+        let(:variant_price) { 15 }
+
+        it { is_expected.to eq("(Add: $5.00)") }
+        # Regression test for https://github.com/spree/spree/issues/2737
+        it { is_expected.to be_html_safe }
+      end
+
+      context "when variant is less than master" do
+        let(:product_price) { 15 }
+
+        it { is_expected.to eq("(Subtract: $5.00)") }
+      end
+    end
+
+    context "when currency is JPY" do
+      let(:variant_price) { 100 }
+      let(:product_price) { 100 }
+      let(:currency) { 'JPY' }
+
+      context "when variant is more than master" do
+        let(:variant_price) { 150 }
+
+        it { is_expected.to eq("(Add: &#x00A5;50)") }
+      end
+
+      context "when variant is less than master" do
+        let(:product_price) { 150 }
+
+        it { is_expected.to eq("(Subtract: &#x00A5;50)") }
+      end
+    end
+  end
+
+  context "#variant_price_full" do
+    before do
+      Spree::Config[:show_variant_full_price] = true
+      @variant1 = create(:variant, :product => product)
+      @variant2 = create(:variant, :product => product)
+    end
+
+    context "when currency is default" do
+      it "should return the variant price if the price is different than master" do
+        product.price = 10
+        @variant1.price = 15
+        @variant2.price = 20
+        expect(helper.variant_price(@variant1)).to eq("$15.00")
+        expect(helper.variant_price(@variant2)).to eq("$20.00")
+      end
+    end
+
+    context "when currency is JPY" do
+      let(:currency) { 'JPY' }
+
+      before do
+        product.variants.active.each do |variant|
+          variant.prices.each do |price|
+            price.currency = currency
+            price.save!
+          end
+        end
+      end
+
+      it "should return the variant price if the price is different than master" do
+        product.price = 100
+        @variant1.price = 150
+        expect(helper.variant_price(@variant1)).to eq("&#x00A5;150")
+      end
+    end
+
+    it "should be nil when all variant prices are equal" do
+      product.price = 10
+      @variant1.default_price.update_column(:amount, 10)
+      @variant2.default_price.update_column(:amount, 10)
+      expect(helper.variant_price(@variant1)).to be_nil
+      expect(helper.variant_price(@variant2)).to be_nil
     end
   end
 end
