+++ frontend/spec/helpers/products_helper_spec.rb
@@ -3,7 +3,7 @@
 require 'spec_helper'
 
 module Spree
-  describe ProductsHelper, :type => :helper do
+  describe ProductsHelper, type: :helper do
     include ProductsHelper
 
     let(:product) { create(:product) }
@@ -12,111 +12,6 @@ module Spree
     before do
       allow(helper).to receive(:current_currency) { currency }
     end
-
-    context "#variant_price_diff" do
-      let(:product_price) { 10 }
-      let(:variant_price) { 10 }
-
-      before do
-        @variant = create(:variant, :product => product)
-        product.price = 15
-        @variant.price = 10
-        allow(product).to receive(:amount_in) { product_price }
-        allow(@variant).to receive(:amount_in) { variant_price }
-      end
-
-      subject { helper.variant_price(@variant) }
-
-      context "when variant is same as master" do
-        it { is_expected.to be_nil }
-      end
-
-      context "when the master has no price" do
-        let(:product_price) { nil }
-
-        it { is_expected.to be_nil }
-      end
-
-      context "when currency is default" do
-        context "when variant is more than master" do
-          let(:variant_price) { 15 }
-
-          it { is_expected.to eq("(Add: $5.00)") }
-          # Regression test for https://github.com/spree/spree/issues/2737
-          it { is_expected.to be_html_safe }
-        end
-
-        context "when variant is less than master" do
-          let(:product_price) { 15 }
-
-          it { is_expected.to eq("(Subtract: $5.00)") }
-        end
-      end
-
-      context "when currency is JPY" do
-        let(:variant_price) { 100 }
-        let(:product_price) { 100 }
-        let(:currency) { 'JPY' }
-
-        context "when variant is more than master" do
-          let(:variant_price) { 150 }
-
-          it { is_expected.to eq("(Add: &#x00A5;50)") }
-        end
-
-        context "when variant is less than master" do
-          let(:product_price) { 150 }
-
-          it { is_expected.to eq("(Subtract: &#x00A5;50)") }
-        end
-      end
-    end
-
-    context "#variant_price_full" do
-      before do
-        Spree::Config[:show_variant_full_price] = true
-        @variant1 = create(:variant, :product => product)
-        @variant2 = create(:variant, :product => product)
-      end
-
-      context "when currency is default" do
-        it "should return the variant price if the price is different than master" do
-          product.price = 10
-          @variant1.price = 15
-          @variant2.price = 20
-          expect(helper.variant_price(@variant1)).to eq("$15.00")
-          expect(helper.variant_price(@variant2)).to eq("$20.00")
-        end
-      end
-
-      context "when currency is JPY" do
-        let(:currency) { 'JPY' }
-
-        before do
-          product.variants.active.each do |variant|
-            variant.prices.each do |price|
-              price.currency = currency
-              price.save!
-            end
-          end
-        end
-
-        it "should return the variant price if the price is different than master" do
-          product.price = 100
-          @variant1.price = 150
-          expect(helper.variant_price(@variant1)).to eq("&#x00A5;150")
-        end
-      end
-
-      it "should be nil when all variant prices are equal" do
-        product.price = 10
-        @variant1.default_price.update_column(:amount, 10)
-        @variant2.default_price.update_column(:amount, 10)
-        expect(helper.variant_price(@variant1)).to be_nil
-        expect(helper.variant_price(@variant2)).to be_nil
-      end
-    end
-
 
     context "#product_description" do
       # Regression test for https://github.com/spree/spree/issues/1607
@@ -180,7 +75,7 @@ THIS IS THE BEST PRODUCT EVER!
       subject { helper.cache_key_for_products }
       before(:each) do
         @products = double('products collection')
-        allow(helper).to receive(:params) { {:page => 10} }
+        allow(helper).to receive(:params) { {page: 10} }
       end
 
       context 'when there is a maximum updated date' do
@@ -203,6 +98,13 @@ THIS IS THE BEST PRODUCT EVER!
 
         it { is_expected.to eq('en/USD/spree/products/all-10-20131211-1234567') }
       end
+    end
+
+    # Regression test for https://github.com/spree/spree/issues/2518 and https://github.com/spree/spree/issues/2323
+    it "truncates HTML correctly in product description" do
+      product = double(description: "<strong>" + ("a" * 95) + "</strong> This content is invisible.")
+      expected = "<strong>" + ("a" * 95) + "</strong>..."
+      expect(truncated_product_description(product)).to eq(expected)
     end
   end
 end
